Update discount to 2.0.7
2.0.7 completes support for definition lists in PHP mardown extra style.
There are also a few small cleanups and an extension of the footnotes.

Also:

Build discount with both types of definition list

Discount now supports two types of definition lists: its own and one
from PHP markdown extra. However, to enable them, we need to change
the configuration options.

See the homepage for details on the two types of definition lists:
http://www.pell.portland.or.us/~orc/Code/discount/#dl

Signed-off-by: Adam Vandenberg <flangy@gmail.com>