sphinx: enable libstemmer support
The upstream libstemmer project constantly changes the contents of the
tarball, so doing checksum verification isn't feasible unless we want to
be updating the formula constantly. Thus, use a head-only sub-formula
without a checksum.

Rather than make it an option, I've just set it to always build with
libstemmer per #5730. We can always add the option later if it's
necessary.

Includes an unrelated fix to put the var directory in
HOMEBREW_PREFIX/var.

Signed-off-by: Jack Nagel <jacknagel@gmail.com>