libmms 0.6.2
This commit updates libmms to version 0.6.2.  Their project is
back at sf.net for hosting, and that's why the change in both url
and homepage. This compiles with llvm-2335, gcc-4.2.1, and clang2.0
on OSX 10.6.8 with XCode 4.0.2. It builds native static and shared
libraries that are used for streaming mmst amd mmsh protocol as in
    mms://i.canhaz.ct:1755
Closes #7991.

Signed-off-by: Jack Nagel <jacknagel@gmail.com>