gpgme: disable gnupg2-only options
Currently, gpgme will build against gnupg2 but this results in a lot of
test failures, at least for me. And if both gnupg and gnupg2 are
installed, the test suite will run against a combination of binaries
from the two packages, resulting in even more failed tests.

So for now, let's keep the dependency on gnupg 1.x only and disable the
gpgme components that require gnupg2.

Patches welcome to fix the underlying issues; I will probably mess
around with it in the future, too.

Signed-off-by: Jack Nagel <jacknagel@gmail.com>