gnupg2: specify path to gpg-agent
We disable building the agent program and provide it as a separate
gpg-agent package so that gnupg 1.x can use it as well. However, gpg2
still tries to find the agent in its own keg if it isn't already
running; that is, if the user hasn't done something like

    eval $(gpg-agent --daemon)

Using --with-agent-pgm, we can tell gpg2 to look in HOMEBREW_PREFIX/bin
for the agent instead.

Signed-off-by: Jack Nagel <jacknagel@gmail.com>